refactoring blueprint and remove cycl dep
