+++ src/cryptoadvance/specterext/spectrum/static/spectrum/css/styles.css
