+++ src/cryptoadvance/specterext/spectrum/templates/spectrum/base.jinja
