+++ src/cryptoadvance/specterext/spectrum/templates/spectrum/components/spectrum_info.jinja
