+++ src/cryptoadvance/specterext/spectrum/templates/spectrum/components/spectrum_menu.jinja
