+++ src/cryptoadvance/specterext/spectrum/templates/spectrum/components/spectrum_tab.jinja
