+++ src/cryptoadvance/specterext/spectrum/templates/spectrum/index.jinja
