+++ src/cryptoadvance/specterext/spectrum/templates/spectrum/settings.jinja
