+++ src/cryptoadvance/specterext/spectrum/templates/spectrum/spectrum_setup.jinja
