+++ src/cryptoadvance/specterext/spectrum/templates/spectrum/welcome/components/get_started.jinja
